backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
sqlalchemy==2.0.48
passlib[bcrypt]==1.7.4
PyJWT==2.12.1
python-multipart==0.0.22
psycopg[binary]==3.2.12
